+++ spec/logging_spec.clj
@@ -1,6 +1,7 @@
 (ns logging-spec
   (:use speclj.core
-        io.aviso.logging.correlation)
+        io.aviso.logging.correlation
+        io.aviso.logging.mdc)
   (:require io.aviso.logging.setup
             [clojure.tools.logging :as l])
   (:import [ch.qos.logback.classic.util ContextSelectorStaticBinder]
@@ -8,10 +9,12 @@
 
 
 (defn- echo-handler
-  [request]
+  [_request]
   {:status  200
    :headers {}
    :body    {:cid *correlation-id*}})
+
+(l/info "Forcing logging initialization ....")
 
 (describe "request correlation"
 
@@ -39,23 +42,67 @@
                   :body    {:cid "override"}}
                  response))))
 
-(describe "MDC"
+
+
+(defmacro capturing-mdc
+  [sym & body]
+  `(do
+     (let [~sym (atom nil)
+           logger-root# (-> (ContextSelectorStaticBinder/getSingleton) .getContextSelector .getLoggerContext (.getLogger "ROOT"))
+           ;; Haven't yet found a way to capture the formatted log output text, but this is pretty
+           ;; safe: it's in the MDC ready to formatted.  Also a user can see it in the logs from the output
+           capture-appender# (reify Appender
+                               (doAppend [_ event#]
+                                 (reset! ~sym
+                                         (into {} (.getMDCPropertyMap event#)))))]
+       (try
+         (.addAppender logger-root# capture-appender#)
+         ~@body
+         (finally
+           (.detachAppender logger-root# capture-appender#))))))
+
+(describe "correlation MDC"
 
   (it "can incorporate the correlation-id into the MDC"
-      (l/info "Forcing init of logging")
-      (let [logged-text (atom nil)
-            logger-root      (-> (ContextSelectorStaticBinder/getSingleton) .getContextSelector .getLoggerContext (.getLogger "ROOT"))
-            ;; Haven't yet found a way to capture the formatted log output text, but this is pretty
-            ;; safe: it's in the MDC ready to formatted.  Also a user can see it in the logs from the output
-            capture-appender (reify Appender
-                               (doAppend [_ event]
-                                 (reset! logged-text
-                                         (-> event .getMDCPropertyMap (.get "correlation-id")))))]
-        (try
-          (.addAppender logger-root capture-appender)
-          (binding [*correlation-id* "XYZ"]
-            (l/info "logging event") (should-contain "XYZ" @logged-text))
-          (finally
-            (.detachAppender logger-root capture-appender))))))
+      (capturing-mdc mdc
+        (binding [*correlation-id* "XYZ"]
+          (l/info "logging event")
+          ;; Due to leakiness in the MDC, a re-run of this test may
+          ;; still include "magic" => "ring".
+          (should= "XYZ" (get @mdc "correlation-id"))))))
+
+(describe "extra MDC"
+  (it "can incorporate extra MDC"
+      (capturing-mdc mdc
+        (with-mdc "magic" "ring"
+                  (l/info "extra mdc logging event")
+                  (should= "ring" (get @mdc "magic")))))
+
+  (it "ignores nil values"
+      (capturing-mdc mdc
+        (with-mdc ["nothing" nil]
+                  (l/info "nothing logging")
+                  (should= false
+                           (contains? @mdc "nothing")))))
+
+  (it "prevents leakage when defaults are available"
+      (set-mdc-default "leaky" "default")
+      (capturing-mdc mdc
+        (l/info "checking default")
+        (should= "default"
+                 (get @mdc "leaky")))
+
+      (capturing-mdc mdc
+        (with-mdc "leaky" "override"
+                  (l/info "inside with-mdc")
+                  (should= {"leaky" "override"}
+                           *extra-mdc*)
+                  (should= "override"
+                           (get @mdc "leaky"))))
+
+      (capturing-mdc mdc
+        (l/info "checking default restored")
+        (should= "default"
+                 (get @mdc "leaky")))))
 
 (run-specs)