+++ CHANGES.md
@@ -1,3 +1,8 @@
+## 0.2.0 - UNRELEASED
+
+New namespace: io.aviso.logging.mdc; a more generalized approach that allows
+arbitrary keys from the `*extra-mdc*` var to be added to the event MDC.
+
 ## 0.1.0 -- 2 Aug 2015
 
 Initial release.